+++ CHANGES.md
@@ -1,5 +1,10 @@
 CHANGES
 =======
+Version 1.0.1 - 2026-08-04
+------------------
+- Black hole animation only
+- Fixing ghost
+
 2026-08-02 -> 1.0.0
 --------------------------------
 * Adapted to back refactoring for first version